requirements.txt
Django==2.2.12
django-crispy-forms==1.10.0
djangorestframework==3.10.3
pytz==2020.4
sqlparse==0.4.1
